+++ FitCloudKit/FitCloudKit.xcframework/ios-arm64/FitCloudKit.framework/Headers/FitCloudKit.h
@@ -10,7 +10,7 @@
 //          FitCloudPro 智能手表 iOS 框架，封装了与手表设备通信等核心功能。
 //
 //  构建版本：
-//      pcjbird    2026-01-28  Version:1.3.2-beta.45 Build:20260128002
+//      pcjbird    2026-02-01  Version:1.3.2-beta.46 Build:20260201001
 
 #import <Foundation/Foundation.h>
 #import <UIKit/UIKit.h>